action mcuboot: Cleanup mcuboot compilation reusable workflow
Stop using composite action to store MCUboot in second bank and do it
locally.
The composite action would be useful if it was possible to iterate over
a step in github action. (It was written with that idea but it turns out
it is not possible).
an hex file is still produced with the bootloader stored in each bank.